# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Sistema de Controle de Visitantes
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: role
          schema:
            type: string
            enum: [admin, receptionist, attendant]
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset user password to the initial password
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Sectors ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ── Auth ──────────────────────────────────────────────────────────────
...
    UserPermissions:
      type: object
      required:
        [
          viewDashboard,
          viewDashboardCharts,
          viewVisits,
          registerVisit,
          checkoutVisit,
          editVisit,
          cancelVisit,
          reprintLabel,
          viewVisitors,
          createVisitor,
          editVisitorName,
          editVisitorCpf,
          editVisitorBirthDate,
          editVisitorPhone,
          editVisitorCompany,
          editVisitorCity,
          accessServiceCenter,
          viewReports,
          manageSectors,
          manageSettings,
          viewAudit,
        ]
      properties:
        viewDashboard:
          type: boolean
        viewDashboardCharts:
          type: boolean
        viewVisits:
          type: boolean
        registerVisit:
          type: boolean
        checkoutVisit:
          type: boolean
        editVisit:
          type: boolean
        cancelVisit:
          type: boolean
        reprintLabel:
          type: boolean
        viewVisitors:
          type: boolean
        createVisitor:
          type: boolean
        editVisitorName:
          type: boolean
        editVisitorCpf:
          type: boolean
        editVisitorBirthDate:
          type: boolean
        editVisitorPhone:
          type: boolean
        editVisitorCompany:
          type: boolean
        editVisitorCity:
          type: boolean
        accessServiceCenter:
          type: boolean
        viewReports:
          type: boolean
        manageSectors:
          type: boolean
        manageSettings:
          type: boolean
        viewAudit:
          type: boolean
    User:
      type: object
      required:
        [
          id,
          name,
          login,
          role,
          status,
          mustChangePassword,
          permissions,
          createdAt,
        ]
      properties:
        id:
          type: integer
        name:
          type: string
        login:
          type: string
        role:
          type: string
          enum: [admin, receptionist, attendant]
        sectorId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, inactive]
        mustChangePassword:
          type: boolean
        permissions:
          $ref: "#/components/schemas/UserPermissions"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    UserInput:
      type: object
      required: [name, login, role]
      properties:
        name:
          type: string
        login:
          type: string
        role:
          type: string
          enum: [admin, receptionist, attendant]
        sectorId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, inactive]
          default: active
        permissions:
          $ref: "#/components/schemas/UserPermissions"
...
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        login:
          type: string
        role:
          type: string
          enum: [admin, receptionist, attendant]
        sectorId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, inactive]
        permissions:
          $ref: "#/components/schemas/UserPermissions"

    # ── Sector ────────────────────────────────────────────────────────────
...